Extract project component to external .tsx file
